Grow mlBuffer on compileJump only if necessary
This can avoid an unnecessary allocation, since grow() unconditionally
grows the buffer, and we don't need it to if the current buffer is
already big enough for the jump.

This is particularly relevant when freeDirective < 0, when the point of
the jump is just to skip the remaining free space. The extra space that
gets allocated in compileJump then gets immediately sliced out anyway,
so you get the same end result after doing a lot more work.